extract time module
