+++ srt-protocol/src/protocol/time.rs
@@ -1,14 +1,156 @@
-use std::cmp::max;
+use std::cmp::{max, Ordering};
+use std::fmt;
+use std::num::Wrapping;
+use std::ops::{Add, Div, Mul, Neg, Sub};
 use std::time::{Duration, Instant};
+use std::u32;
 
-/// Timestamp in us
-pub type TimeStamp = i32;
+/// Timestamp in us after creation
+/// These wrap every 2^32 microseconds
+#[derive(Copy, Clone, PartialEq, Eq, Ord)]
+pub struct TimeStamp(Wrapping<u32>);
 
-/// Duration in us, e.g. RTT
-pub type TimeSpan = i32;
+/// Signed duration in us, e.g. RTT
+#[derive(Copy, Clone, PartialEq, Eq, Ord, PartialOrd)]
+pub struct TimeSpan(i32);
 
 #[derive(Copy, Clone, Debug)]
 pub struct TimeBase(Instant);
+
+impl TimeStamp {
+    pub fn from_micros(us: u32) -> Self {
+        Self(Wrapping(us))
+    }
+
+    pub fn as_micros(self) -> u32 {
+        (self.0).0
+    }
+}
+
+impl fmt::Debug for TimeStamp {
+    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+        let time = (self.0).0;
+        let mins = time / 1_000_000 / 60 % 60;
+        let secs = time / 1_000_000 % 60;
+        let micros = time % 1_000_000;
+
+        write!(f, "{:02}:{:02}.{:06}", mins, secs, micros)
+    }
+}
+
+impl PartialOrd<TimeStamp> for TimeStamp {
+    fn partial_cmp(&self, other: &TimeStamp) -> Option<Ordering> {
+        // this is a "best effort" implementation, and goes for close
+        // if timestamps are very far apart, this will not work (and cannot)
+        Some((*self - *other).as_micros().cmp(&0))
+    }
+}
+
+impl Add<TimeSpan> for TimeStamp {
+    type Output = TimeStamp;
+
+    #[allow(clippy::suspicious_arithmetic_impl)]
+    fn add(self, rhs: TimeSpan) -> Self::Output {
+        TimeStamp(if rhs.0 > 0 {
+            self.0 + Wrapping(rhs.0 as u32)
+        } else {
+            self.0 - Wrapping(rhs.0.abs() as u32)
+        })
+    }
+}
+
+impl Sub<TimeSpan> for TimeStamp {
+    type Output = TimeStamp;
+
+    fn sub(self, rhs: TimeSpan) -> Self::Output {
+        self + -rhs
+    }
+}
+
+impl Sub<TimeStamp> for TimeStamp {
+    type Output = TimeSpan;
+
+    fn sub(self, rhs: TimeStamp) -> TimeSpan {
+        // This is also a "best effort" implementation, and cannot be precise
+        let pos_sub = self.0 - rhs.0;
+        let neg_sub = rhs.0 - self.0;
+        if pos_sub < neg_sub {
+            TimeSpan(pos_sub.0 as i32)
+        } else {
+            -TimeSpan(neg_sub.0 as i32)
+        }
+    }
+}
+
+impl TimeSpan {
+    pub fn from_micros(us: i32) -> Self {
+        Self(us)
+    }
+
+    pub fn as_micros(self) -> i32 {
+        self.0
+    }
+
+    pub fn abs(self) -> Self {
+        Self(self.0.abs())
+    }
+
+    pub fn as_secs_f64(self) -> f64 {
+        self.0 as f64 / 1e6
+    }
+}
+
+impl fmt::Debug for TimeSpan {
+    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+        let sign = if self.0 < 0 { "-" } else { "" };
+        let mins = self.0.abs() / 1_000_000 / 60 % 60;
+        let secs = self.0.abs() / 1_000_000 % 60;
+        let micros = self.0.abs() % 1_000_000;
+
+        write!(f, "{}{:02}:{:02}.{:06}", sign, mins, secs, micros)
+    }
+}
+
+impl Neg for TimeSpan {
+    type Output = TimeSpan;
+
+    fn neg(self) -> Self::Output {
+        Self(-self.0)
+    }
+}
+
+impl Mul<i32> for TimeSpan {
+    type Output = TimeSpan;
+
+    fn mul(self, rhs: i32) -> Self::Output {
+        Self(self.0 * rhs)
+    }
+}
+
+impl Add<TimeSpan> for TimeSpan {
+    type Output = TimeSpan;
+
+    fn add(self, rhs: TimeSpan) -> Self::Output {
+        Self(self.0 + rhs.0)
+    }
+}
+
+impl Div<i32> for TimeSpan {
+    type Output = TimeSpan;
+
+    fn div(self, rhs: i32) -> Self::Output {
+        Self(self.0 / rhs)
+    }
+}
+
+impl Sub<TimeSpan> for TimeSpan {
+    type Output = TimeSpan;
+
+    fn sub(self, rhs: TimeSpan) -> Self::Output {
+        Self(self.0 - rhs.0)
+    }
+}
+
 impl TimeBase {
     pub fn new(start_time: Instant) -> Self {
         Self(start_time)
@@ -16,32 +158,79 @@ impl TimeBase {
 
     #[allow(clippy::trivially_copy_pass_by_ref)]
     pub fn timestamp_from(&self, instant: Instant) -> TimeStamp {
-        if self.0 > instant {
-            -((self.0 - instant).as_micros() as i32)
-        } else {
-            (instant - self.0).as_micros() as i32
-        }
+        const TIMESTAMP_MASK: u128 = u32::MAX as u128;
+
+        assert!(
+            self.0 <= instant,
+            "Timestamps are only valid after the timebase start time"
+        );
+        TimeStamp(Wrapping(
+            ((instant - self.0).as_micros() & TIMESTAMP_MASK) as u32,
+        ))
     }
 
+    // Get Instant closest to `now` that is consistent with `timestamp`
     #[allow(clippy::trivially_copy_pass_by_ref)]
-    pub fn instant_from(&self, timestamp: TimeStamp) -> Instant {
-        if timestamp < 0 {
-            self.0 - Duration::from_micros(timestamp.abs() as u64)
-        } else {
-            self.0 + Duration::from_micros(timestamp as u64)
-        }
+    pub fn instant_from(&self, now: Instant, timestamp: TimeStamp) -> Instant {
+        let wraps = ((now - self.0).as_micros() >> 32) as u64;
+        self.0
+            + Duration::from_micros(wraps * u64::from(std::u32::MAX) + timestamp.as_micros() as u64)
     }
 
     pub fn adjust(&mut self, delta: TimeSpan) {
-        match delta {
-            delta if delta > 0 => {
-                self.0 += Duration::from_micros(delta as u64);
-            }
-            delta if delta < 0 => {
-                self.0 -= Duration::from_micros(delta.abs() as u64);
-            }
-            _ => {}
+        if delta.0 > 0 {
+            self.0 += Duration::from_micros(delta.0 as u64);
+        } else {
+            self.0 -= Duration::from_micros(delta.0.abs() as u64);
         }
+    }
+
+    pub fn origin_time(&self) -> Instant {
+        self.0
+    }
+}
+
+#[cfg(test)]
+mod timestamp {
+    use super::*;
+
+    #[test]
+    #[allow(clippy::eq_op)]
+    fn subtract_timestamp() {
+        let a = TimeStamp::from_micros(10);
+        let max = a - TimeSpan(11);
+        let b = TimeStamp::from_micros(11);
+
+        assert_eq!(a - a, TimeSpan::from_micros(0));
+        assert_eq!(b - a, TimeSpan::from_micros(1));
+        assert_eq!(a - b, TimeSpan::from_micros(-1));
+        assert!(max < a);
+        assert!(b > a);
+        assert!(b > max);
+        assert_eq!(max.as_micros(), u32::MAX);
+    }
+
+    #[test]
+    fn debug_fmt() {
+        assert_eq!("00:00.000001", format!("{:?}", TimeStamp::from_micros(1)));
+        assert_eq!(
+            "01:02.030040",
+            format!("{:?}", TimeStamp::from_micros(62030040))
+        );
+    }
+}
+
+#[cfg(test)]
+mod timespan {
+    use super::*;
+
+    #[test]
+    fn debug_fmt() {
+        assert_eq!("00:00.000001", format!("{:?}", TimeSpan::from_micros(1)));
+        assert_eq!(
+            "-01:02.030040",
+            format!("{:?}", TimeSpan::from_micros(-62030040))
+        );
     }
 }
 
@@ -52,83 +241,39 @@ mod timebase {
 
     proptest! {
         #[test]
-        fn timestamp_roundtrip(expected_ts: i32) {
+        fn timestamp_roundtrip(expected_ts: u32) {
             let timebase = TimeBase::new(Instant::now());
+            let expected_ts = TimeStamp::from_micros(expected_ts);
 
-            let ts = timebase.timestamp_from(timebase.instant_from(expected_ts));
-            prop_assert_eq!(ts, expected_ts);
+            let ts = timebase.timestamp_from(timebase.instant_from(Instant::now(), expected_ts));
+            assert_eq!(ts, expected_ts);
         }
 
         #[test]
-        fn timestamp_from(expected_ts: i32, time_domain in -5i64..5) {
-            let micros = |micros| Duration::from_micros(micros as u64);
+        fn timestamp_from(expected_ts: u32, n in 0u64..10) {
             let now = Instant::now();
             let timebase = TimeBase::new(now);
-
-            let rollover_interval = std::u32::MAX as i64 + 1;
-            let delta = time_domain * rollover_interval + expected_ts as i64;
-            let instant = if delta > 0 { now + micros(delta) } else { now - micros(delta.abs()) };
-
+            let delta = ((std::u32::MAX as u64 + 1)* n) + expected_ts as u64;
+            let instant =  now + Duration::from_micros(delta as u64);
             let ts = timebase.timestamp_from(instant);
-
-            prop_assert_eq!(ts, expected_ts);
+            assert_eq!(ts, TimeStamp::from_micros(expected_ts));
         }
 
         #[test]
         fn adjust(drift: i16) {
             let now = Instant::now();
             let mut timebase = TimeBase::new(now);
+            let drift = TimeSpan::from_micros(i32::from(drift));
 
             let original_ts = timebase.timestamp_from(now);
-            timebase.adjust(drift as TimeSpan);
-            let ts = timebase.timestamp_from(now);
+            timebase.adjust(drift);
+            let ts = timebase.timestamp_from(now + Duration::from_micros(1_000_000));
 
-            prop_assert_eq!(ts, original_ts - drift as TimeSpan);
+            assert_eq!(ts, original_ts - drift + TimeSpan::from_micros(1_000_000));
         }
+
     }
 }
-
-//4. Timers
-//
-//   UDT uses four timers to trigger different periodical events. Each
-//   event has its own period and they are all independent. They use the
-//   system time as origins and should process wrapping if the system time
-//   wraps.
-//
-//   For a certain periodical event E in UDT, suppose the time variable is
-//   ET and its period is p. If E is set or reset at system time t0 (ET =
-//   t0), then at any time t1, (t1 - ET >= p) is the condition to check if
-//   E should be triggered.
-//
-//   The four timers are ACK, NAK, EXP and SND. SND is used in the sender
-//   only for rate-based packet sending (see Section 6.1), whereas the
-//   other three are used in the receiver only.
-//
-//   ACK is used to trigger an acknowledgement (ACK). Its period is set by
-//   the congestion control module. However, UDT will send an ACK no
-//   longer than every 0.01 second, even though the congestion control
-//   does not need timer-based ACK. Here, 0.01 second is defined as the
-//   SYN time, or synchronization time, and it affects many of the other
-//   timers used in UDT.
-//
-//   NAK is used to trigger a negative acknowledgement (NAK). Its period
-//   is dynamically updated to 4 * RTT_+ RTTVar + SYN, where RTTVar is the
-//   variance of RTT samples.
-//
-//   EXP is used to trigger data packets retransmission and maintain
-//   connection status. Its period is dynamically updated to 4 * RTT +
-//   RTTVar + SYN.
-//
-//   The recommended granularity of their periods is microseconds. The
-//   system time is queried after each time bounded UDP receiving (there
-//   will be additional necessary data processing time if a UDP packet is
-//   received) to check if any of the ACK, NAK, or EXP event should be
-//   triggered. The timeout value of UDP receiving should be at least SYN.
-//
-//   In the rest of this document, a name of a time variable will be used
-//   to represent the associated event, the variable itself, or the value
-//   of its period, depending on the context. For example, ACK can mean
-//   either the ACK event or the value of ACK period.
 
 pub struct Timer {
     period: Duration,
@@ -143,6 +288,10 @@ impl Timer {
             period: max(period, Self::MIN_PERIOD),
             last: now,
         }
+    }
+
+    pub fn period(&mut self) -> Duration {
+        self.period
     }
 
     pub fn next_instant(&self) -> Instant {
